+++ packages/mistralai-gcp/src/models/components/chatcompletionstreamrequest.ts
@@ -78,7 +78,7 @@ export type ChatCompletionStreamRequest = {
   /**
    * ID of the model to use. You can use the [List Available Models](/api/#tag/models/operation/list_models_v1_models_get) API to see all of your available models, or see our [Model overview](/models) for model descriptions.
    */
-  model: string | null;
+  model: string;
   /**
    * What sampling temperature to use, we recommend between 0.0 and 0.7. Higher values like 0.7 will make the output more random, while lower values like 0.2 will make it more focused and deterministic. We generally recommend altering this or `top_p` but not both. The default value varies depending on the model you are targeting. Call the `/models` endpoint to retrieve the appropriate value.
    */
@@ -125,6 +125,7 @@ export type ChatCompletionStreamRequest = {
    */
   n?: number | null | undefined;
   prediction?: Prediction | undefined;
+  parallelToolCalls?: boolean | undefined;
 };
 
 /** @internal */
@@ -308,7 +309,7 @@ export const ChatCompletionStreamRequest$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  model: z.nullable(z.string()),
+  model: z.string(),
   temperature: z.nullable(z.number()).optional(),
   top_p: z.number().optional(),
   max_tokens: z.nullable(z.number().int()).optional(),
@@ -347,6 +348,7 @@ export const ChatCompletionStreamRequest$inboundSchema: z.ZodType<
   frequency_penalty: z.number().optional(),
   n: z.nullable(z.number().int()).optional(),
   prediction: Prediction$inboundSchema.optional(),
+  parallel_tool_calls: z.boolean().optional(),
 }).transform((v) => {
   return remap$(v, {
     "top_p": "topP",
@@ -356,12 +358,13 @@ export const ChatCompletionStreamRequest$inboundSchema: z.ZodType<
     "tool_choice": "toolChoice",
     "presence_penalty": "presencePenalty",
     "frequency_penalty": "frequencyPenalty",
+    "parallel_tool_calls": "parallelToolCalls",
   });
 });
 
 /** @internal */
 export type ChatCompletionStreamRequest$Outbound = {
-  model: string | null;
+  model: string;
   temperature?: number | null | undefined;
   top_p?: number | undefined;
   max_tokens?: number | null | undefined;
@@ -381,6 +384,7 @@ export type ChatCompletionStreamRequest$Outbound = {
   frequency_penalty?: number | undefined;
   n?: number | null | undefined;
   prediction?: Prediction$Outbound | undefined;
+  parallel_tool_calls?: boolean | undefined;
 };
 
 /** @internal */
@@ -389,7 +393,7 @@ export const ChatCompletionStreamRequest$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ChatCompletionStreamRequest
 > = z.object({
-  model: z.nullable(z.string()),
+  model: z.string(),
   temperature: z.nullable(z.number()).optional(),
   topP: z.number().optional(),
   maxTokens: z.nullable(z.number().int()).optional(),
@@ -430,6 +434,7 @@ export const ChatCompletionStreamRequest$outboundSchema: z.ZodType<
   frequencyPenalty: z.number().optional(),
   n: z.nullable(z.number().int()).optional(),
   prediction: Prediction$outboundSchema.optional(),
+  parallelToolCalls: z.boolean().optional(),
 }).transform((v) => {
   return remap$(v, {
     topP: "top_p",
@@ -439,6 +444,7 @@ export const ChatCompletionStreamRequest$outboundSchema: z.ZodType<
     toolChoice: "tool_choice",
     presencePenalty: "presence_penalty",
     frequencyPenalty: "frequency_penalty",
+    parallelToolCalls: "parallel_tool_calls",
   });
 });