+++ packages/mistralai-gcp/RELEASES.md
@@ -109,3 +109,13 @@ Based on:
 - [typescript v1.4.0] packages/mistralai-gcp
 ### Releases
 - [NPM v1.4.0] https://www.npmjs.com/package/@mistralai/mistralai-gcp/v/1.4.0 - packages/mistralai-gcp
+
+## 2025-03-20 14:22:31
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.517.3 (2.548.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.5.0] packages/mistralai-gcp
+### Releases
+- [NPM v1.5.0] https://www.npmjs.com/package/@mistralai/mistralai-gcp/v/1.5.0 - packages/mistralai-gcp